Started refactor attempt
