+++ bamboo/__init__.py
@@ -1,4 +1,2 @@
 from .main import *
-import bamboo.cooling
-import bamboo.plot
-import bamboo.materials
+import bamboo.compressible
